requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
gitpython>=3.1.40
rich>=13.7.0
typer>=0.9.0 